Replace template files with Ada code to generate text files
We already have manually-written Ada code to generate text files
(Alr.Templates), so it's not a big deal. On the other hand, this removes
the need to clone the whole alire repository just for a handful of text
files: this will make it easier to test alr/alire and will improve
general user experience.